requirements.txt
openai>=1.86.0
python-dotenv>=1.1.0
rich>=14.0.0
elevenlabs>=2.3.0 